website/server/requirements.txt
Django==3.1.14
mysqlclient==2.0.1
django-mysql==3.7.1
django-allauth==0.45.0
django-redis-sessions==0.6.1
djangorestframework==3.11.2
django-cors-headers==3.5.0
django-rest-auth[with_social]==0.9.5
djangorestframework-api-key==2.0.0
PyJWT==2.4.0
djangorestframework-simplejwt==4.8.0
django-prometheus==2.1.0
redis==3.5.2
gunicorn==20.0.4
requests==2.31.0
Pillow==9.3.0
livekit-api